MarkovChainName_parallel.py Updated to V0.2.
Now uses a "push to queue" system to deal with name generation. This
seems to improve run time a lot; close to serial for small sets.
However, it still under performs the serial code, and so will not yet
replace it.

Default job number changed to number_of_cpus * 1.5, as this gives the
best results on my system.